refactor: lot of stuff

+++ Labo1/add.cpp
@@ -9,6 +9,6 @@ Compilateur     : Mingw-w64 g++ 8.1.0
 */
 
 #include "add.h"
-long long Add::apply(long long a, long long b){
-    return a + b;
+long long Add::apply(unsigned a, unsigned b){
+    return (long long) a + (long long) b;
 }
