+++ Labo1/multiply.h
@@ -16,7 +16,7 @@ Compilateur     : Mingw-w64 g++ 8.1.0
 
 class Multiply : public Operation{
 public:
-    long long apply(long long a, long long b);
+    long long apply(unsigned a, unsigned b);
 };
