+++ Labo1/substract.cpp
@@ -9,6 +9,6 @@ Compilateur     : Mingw-w64 g++ 8.1.0
 */
 
 #include "substract.h"
-long long Substract::apply(long long a, long long b){
-    return a - b;
+long long Substract::apply(unsigned a, unsigned b){
+    return (long long) a - (long long) b;
 }
